Move ControllableAction to testtools
